Update index.html
Chrome for Android hides the address bar as you scroll down. This fires a resize event due to the value of window.innerHeight changing. This became problematic because my onresize callback re-renders the content on the page. So I've added a condition  to the resize callback that checks the value of document.documentElement.clientHeight against it's previous value.